добалена стартовая разметка, роутинг
