--- crm/src/App.css
@@ -1,4 +1,0 @@
-.employees {
-    display: flex;
-    gap: 20px;
-}
