--- crm/src/components/EmployeeCard/index.ts
@@ -1,1 +1,0 @@
-export * from './employeeCard';
