+++ crm/src/pages/Customers/CustomersPage.tsx
@@ -1,0 +1,5 @@
+export function CustomersPage() {
+    return (
+        <h1>Клиенты</h1>
+    )
+}
