+++ crm/src/pages/Customers/index.ts
@@ -1,0 +1,1 @@
+export * from './CustomersPage'
