+++ crm/src/pages/Employee/index.ts
@@ -1,0 +1,1 @@
+export * from './components/EmployeeCard';
