+++ crm/src/pages/Home/HomePage.tsx
@@ -1,0 +1,5 @@
+export function HomePage() {
+    return (
+        <h1>Домашняя страница</h1>
+    )
+}
