+++ crm/src/pages/Login/LoginPage.tsx
@@ -1,0 +1,5 @@
+export function LoginPage() {
+    return (
+        <h1>Логин</h1>
+    )
+}
