+++ crm/src/pages/Orders/OrdersPage.tsx
@@ -1,0 +1,5 @@
+export function OrdersPage() {
+    return (
+        <h1>Заказы</h1>
+    )
+}
