+++ crm/src/pages/Services/ServicesPage.tsx
@@ -1,0 +1,5 @@
+export function ServicesPage() {
+    return (
+        <h1>Услуги</h1>
+    )
+}
